+++ test/authentication/login.js
@@ -25,8 +25,8 @@ describe("mCAP.authentication", function () {
 //      console.log(JSON.stringify(xhr, null, 10));
       var postData = JSON.parse(xhr.requestBody);
 
-      expect(postData.username).toBe('m.mustermann');
-      expect(postData.organization).toBe('org');
+      expect(postData.userName).toBe('m.mustermann');
+      expect(postData.orgaName).toBe('org');
       expect(postData.password).toBe('pass');
 
       if (xhr.method == "POST" && xhr.url === mCAP.authentication.url() + 'login') {
@@ -45,15 +45,15 @@ describe("mCAP.authentication", function () {
 
   it("Login config with set authentication successful", function () {
 
-    mCAP.authentication.set('username', 'm.mustermann');
-    mCAP.authentication.set('organization', 'org');
+    mCAP.authentication.set('userName', 'm.mustermann');
+    mCAP.authentication.set('orgaName', 'org');
     mCAP.authentication.set('password', 'pass');
 
     server.respondWith(serverSuccCallback);
 
     mCAP.authentication.login().then(function (data) {
-      expect(mCAP.authentication.get('username')).toBe('m.mustermann');
-      expect(mCAP.authentication.get('organization')).toBe('org');
+      expect(mCAP.authentication.get('userName')).toBe('m.mustermann');
+      expect(mCAP.authentication.get('orgaName')).toBe('org');
       expect(mCAP.authentication.get('password')).toBe('pass');
       delete data.attributes;
       callback(data);
@@ -71,12 +71,12 @@ describe("mCAP.authentication", function () {
     server.respondWith(serverSuccCallback);
 
     mCAP.authentication.login({
-      username: 'm.mustermann',
-      organization: 'org',
+      userName: 'm.mustermann',
+      orgaName: 'org',
       password: 'pass'
     }).then(function (data) {
-      expect(mCAP.authentication.get('username')).toBe('m.mustermann');
-      expect(mCAP.authentication.get('organization')).toBe('org');
+      expect(mCAP.authentication.get('userName')).toBe('m.mustermann');
+      expect(mCAP.authentication.get('orgaName')).toBe('org');
       expect(mCAP.authentication.get('password')).toBe('pass');
       delete data.attributes;
       callback(data);
@@ -88,18 +88,68 @@ describe("mCAP.authentication", function () {
     sinon.assert.calledOnce(callback);
   });
 
-  it("Login config with set username and organisation successful", function () {
+  it("Login config with set userName and organisation successful", function () {
 
-    mCAP.authentication.set('username', 'm.mustermann');
-    mCAP.authentication.set('organization', 'org');
+    mCAP.authentication.set('userName', 'm.mustermann');
+    mCAP.authentication.set('orgaName', 'org');
 
     server.respondWith(serverSuccCallback);
 
     expect(mCAP.authentication.get('password')).toBe('');
 
     mCAP.authentication.login('pass').then(function (data) {
-      expect(mCAP.authentication.get('username')).toBe('m.mustermann');
-      expect(mCAP.authentication.get('organization')).toBe('org');
+      expect(mCAP.authentication.get('userName')).toBe('m.mustermann');
+      expect(mCAP.authentication.get('orgaName')).toBe('org');
+      expect(mCAP.authentication.get('password')).toBe('');
+      delete data.attributes;
+      callback(data);
+    });
+
+    server.respond();
+
+    sinon.assert.calledWith(callback, loginResponseDataSucc);
+    sinon.assert.calledOnce(callback);
+
+  });
+
+  it("Login user is set", function () {
+
+    mCAP.authentication.set('userName', 'm.mustermann');
+    mCAP.authentication.set('orgaName', 'org');
+
+    server.respondWith(serverSuccCallback);
+
+    expect(mCAP.authentication.get('password')).toBe('');
+
+    mCAP.authentication.login('pass').then(function (data) {
+      expect(mCAP.User.prototype.isPrototypeOf(mCAP.authentication.get('user'))).toBeTruthy();
+      expect(mCAP.authentication.get('userName')).toBe('m.mustermann');
+      expect(mCAP.authentication.get('orgaName')).toBe('org');
+      expect(mCAP.authentication.get('password')).toBe('');
+      delete data.attributes;
+      callback(data);
+    });
+
+    server.respond();
+
+    sinon.assert.calledWith(callback, loginResponseDataSucc);
+    sinon.assert.calledOnce(callback);
+
+  });
+
+  it("Login organisation is set", function () {
+
+    mCAP.authentication.set('userName', 'm.mustermann');
+    mCAP.authentication.set('orgaName', 'org');
+
+    server.respondWith(serverSuccCallback);
+
+    expect(mCAP.authentication.get('password')).toBe('');
+
+    mCAP.authentication.login('pass').then(function (data) {
+      expect(mCAP.Organization.prototype.isPrototypeOf(mCAP.authentication.get('organization'))).toBeTruthy();
+      expect(mCAP.authentication.get('userName')).toBe('m.mustermann');
+      expect(mCAP.authentication.get('orgaName')).toBe('org');
       expect(mCAP.authentication.get('password')).toBe('');
       delete data.attributes;
       callback(data);