+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.3 - 2014-06-25
+- update login api
+- implement organization model
+- implement organizations collection
+- add organization model to authentication object
+
 ## 0.0.2 - 2014-06-24
 - add dist